requirements.txt
altair==4.2.2
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
azure-core==1.30.2
azure-storage-blob==12.20.0
babel==2.16.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.55.0
fqdn==1.5.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
h5py==3.11.0
hmmlearn==0.3.3
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.2
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kagglehub==0.3.4
kiwisolver==1.4.7
logparser3==1.0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.30.1
openpyxl==3.1.2
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
pip==24.2
platformdirs==4.3.6
playwright==1.49.0
pluggy==1.5.0
polars==0.20.31
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.9.0
Pympler==1.1
pyparsing==3.2.0
pyspellchecker==0.8.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.0
referencing==0.35.1
regex==2022.3.2
requests==2.31.0
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.5
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.3.1
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
streamlit-aggrid==1.0.5
streamlit-javascript==0.1.5
sweetviz==2.3.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.3.0
terminado==0.18.1
thefuzz==0.22.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tld==0.13
tldextract==5.1.3
tokenize_rt==6.1.0
tokenizers==0.20.3
toml==0.10.2
toolz==0.12.1
torch==2.5.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==4.3.1
uri-template==1.3.0
urllib3==2.2.1
validators==0.33.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==4.0.13
zipp==3.19.2
